加入redis缓存，增强服务

+++ timebook-service/timebook-uniapp-service/src/main/java/hc/service/impl/AlbumServiceImpl.java
@@ -1,6 +1,7 @@
 package hc.service.impl;
 
 
+import cn.hutool.core.collection.CollUtil;
 import cn.hutool.core.util.StrUtil;
 
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
@@ -20,13 +21,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
-import java.util.ArrayList;
-import java.util.List;
+import java.util.*;
 import java.util.concurrent.TimeUnit;
+import java.util.function.Function;
 
 import static hc.LogUtils.info;
-import static hc.common.constants.RedisConstants.CACHE_ALBUM_KEY;
-import static hc.common.constants.RedisConstants.CACHE_ALBUM_TTL;
+import static hc.common.constants.RedisConstants.*;
 import static hc.common.enums.AppHttpCodeEnum.*;
 
 @Service
@@ -51,7 +51,7 @@ public class AlbumServiceImpl extends ServiceImpl<AlbumMapper, Album> implements
             throw new ParamErrorException(NO_FOUND_PARAM);
         }
         Album album=redisCacheClient.queryWithPassThrough(CACHE_ALBUM_KEY,albumId
-                ,Album.class,this::getById,CACHE_ALBUM_TTL, TimeUnit.MINUTES);
+                ,Album.class,this::getById,CACHE_ALBUM_TTL, TimeUnit.DAYS);
         if(album!=null)
             return true;
         return false;
@@ -61,15 +61,9 @@ public class AlbumServiceImpl extends ServiceImpl<AlbumMapper, Album> implements
     public ResponseResult getAll() {
         List<Album> albumList=query().eq("user_id", UserHolder.getUser().getUserId())
                 .orderByAsc("update_time").list();
-        List<Album> albumResult=new ArrayList<>();
-        Album defaultAlbum=query().eq("user_id", UserHolder.getUser().getUserId())
-                .eq("type",DEFAULT_ALBUM_TYPE).one();
-        albumResult.add(defaultAlbum);
-        for(Album album:albumList){
-            if(album.getType()==DEFAULT_ALBUM_TYPE)
-                continue;
-            albumResult.add(album);
-        }
+        if(CollUtil.isEmpty(albumList))
+            return ResponseResult.okResult(SUCCESS.getCode(),"无数据");
+        List<Album> albumResult=albumSortByDesc(albumList);
         return ResponseResult.okResult(albumResult);
     }
 
@@ -97,7 +91,9 @@ public class AlbumServiceImpl extends ServiceImpl<AlbumMapper, Album> implements
 
     @Override
     public Boolean setCover(String albumId, String imageUrl) {
-        Album album = getById(albumId);
+        Album album=redisCacheClient.queryWithPassThrough(CACHE_ALBUM_KEY,albumId,Album.class,
+                this::getById,CACHE_ALBUM_TTL,TimeUnit.DAYS);
+        stringRedisTemplate.delete(CACHE_ALBUM_KEY+albumId);
         if(album==null){
             info("无法设置封面，不存在该相册");
            return false;
@@ -118,7 +114,8 @@ public class AlbumServiceImpl extends ServiceImpl<AlbumMapper, Album> implements
 
     @Override
     public ResponseResult updateAlbum(String albumId, String albumName) {
-        Album album=getById(albumId);
+        Album album=redisCacheClient.queryWithPassThrough(CACHE_ALBUM_KEY,albumId,Album.class,
+                this::getById,CACHE_ALBUM_TTL,TimeUnit.DAYS);
         if(album==null)
             return ResponseResult.errorResult(DATA_NOT_EXIST,"没有该相册");
         Sensitive sensitive=new Sensitive();
@@ -136,13 +133,69 @@ public class AlbumServiceImpl extends ServiceImpl<AlbumMapper, Album> implements
 
     @Override
     public ResponseResult deleteAlbum(String albumId) {
-        Album album = getById(albumId);
-        if(album.getType()!=DEFAULT_ALBUM_TYPE){
-            removeById(albumId);
-            imageAlbumService.deleteOtherAlbum(albumId);
-            return ResponseResult.okResult(200,"删除成功");
+        if(StrUtil.isBlank(albumId)){
+            info("albumId为空，不可删除");
+            throw new ParamErrorException(NO_FOUND_PARAM);
         }
-        return ResponseResult.errorResult(DATA_EXIST,"默认不可删除");
+        Album album=redisCacheClient.queryWithPassThrough(CACHE_ALBUM_KEY,albumId,Album.class,
+                this::getById,CACHE_ALBUM_TTL,TimeUnit.DAYS);
+        if(album!=null) {
+            if (album.getType() != DEFAULT_ALBUM_TYPE) {
+                stringRedisTemplate.delete(CACHE_ALBUM_KEY + albumId);
+                removeById(albumId);
+                imageAlbumService.deleteOtherAlbum(albumId);
+                return ResponseResult.okResult(200, "删除成功");
+            } else
+                return ResponseResult.errorResult(DATA_EXIST, "默认不可删除");
+        } else {
+            return ResponseResult.errorResult(DATA_NOT_EXIST, "网络出现问题，请稍后访问");
+        }
+    }
+
+    public String getRedisCacheAlbumId(String userId){
+        String key= DEFAULT_ALBUM_KEY+userId;
+        String json=stringRedisTemplate.opsForValue().get(key);
+        //2.判断是否存在
+        if(StrUtil.isNotBlank(json)){
+            //3.存在，直接返回
+            return json;
+        }
+        //判断命中的是否是空值
+        if(json!=null){
+            return null;
+        }
+        //4.不存在，根据id查询数据库
+        Album album=query().eq("user_id",UserHolder.getUser().getUserId())
+                .eq("type",DEFAULT_ALBUM_TYPE).one();
+        if(StrUtil.isNotBlank(album.getAlbumId()))
+            stringRedisTemplate.opsForValue().set(key,album.getAlbumId(),DEFAULT_ALBUM_TTL,TimeUnit.DAYS);
+        else
+            stringRedisTemplate.opsForValue().set(key,"",CACHE_NULL_TTL,TimeUnit.MINUTES);
+        return album.getAlbumId();
+    }
+
+    private List<Album> albumSortByDesc(List<Album> list){
+        if(CollUtil.isEmpty(list))
+            return null;
+        Album defaultAlbum=null;
+        int removeId=0;
+        Collections.sort(list, new Comparator<Album>() {
+            @Override
+            public int compare(Album a1, Album a2) {
+                return a2.getCreateTime().compareTo(a1.getCreateTime());
+            }
+        });
+        for(Album album:list){
+            if(album.getType()==0) {
+                defaultAlbum = album;
+                removeId=list.indexOf(album);
+            }
+        }
+        if(defaultAlbum!=null) {
+            list.remove(removeId);
+            list.add(0, defaultAlbum);
+        }
+        return list;
     }
 
 }
